+++ dandl-tag.sh
@@ -5,13 +5,14 @@ DELAY=1
 RESUME=false
 
 function usage {
-		echo "./$(basename "$0") [-t] [-s] [-r] -a tag -a tag2"
-        echo "Mass downloader for Gelbooru"
+		echo "./$(basename "$0") [-t] [-s] [-r]"
+        echo "Mass downloader for Danbooru"
 		echo "Simply make a files.txt inside a folder and paste all your links, then run this script to download them all!"
         echo "	-h	shows this help message"
 		echo "	-t	downloads using tor (requires torsocks)"
 		echo "	-s	sets the delay after each request, defaults to 1"
 		echo "	-a	tag or artist name"
+		echo "	-r	will download until it hits a file that already exists"
 }
 
 # list of arguments expected in the input
@@ -55,39 +56,47 @@ for TAG in "${TAGS[@]}"; do
 	cd "$TAG" || exit
 	# GET TAG TOTAL COUNT
 	if $USE_TOR; then
-		TAG_COUNT=$(torsocks curl -s "https://gelbooru.com/index.php?page=dapi&s=tag&q=index&json=1&name=$TAG" | jq -r '."tag" | .[] | ."count"')
+		TAG_COUNT=$(torsocks curl -s "https://danbooru.donmai.us/tags.json?only=id,name,post_count&search\[name_matches\]=$TAG" | jq -r '.[].post_count')
 	else
-		TAG_COUNT=$(curl -s "https://gelbooru.com/index.php?page=dapi&s=tag&q=index&json=1&name=$TAG" | jq -r '."tag" | .[] | ."count"')
+		TAG_COUNT=$(curl -s "https://danbooru.donmai.us/tags.json?only=id,name,post_count&search\[name_matches\]=$TAG" | jq -r '.[].post_count')
 	fi
-
 	# NESTED LOOP TO GET ALL POSTS UNDER TAG
-	TAG_PAGES=$((TAG_COUNT / 100))
+	TAG_PAGES=$((TAG_COUNT / 200))
 	for (( PAGE = 0; PAGE <= TAG_PAGES; PAGE++ ))
 	do
 		if $USE_TOR; then
-			JSON+=$(torsocks curl -s "https://gelbooru.com/index.php?page=dapi&s=post&q=index&json=1&pid=$PAGE&tags=$TAG" | jq -r '.post')
+			JSON_URL+=$(torsocks curl -s "https://danbooru.donmai.us/posts.json?page=$PAGE&limit=200&tags=$TAG" | jq -r '.[].id')
+			if ((PAGE < TAG_PAGES)); then
+				JSON_URL+=$'\n'
+			fi
 		else
-			JSON+=$(curl -s "https://gelbooru.com/index.php?page=dapi&s=post&q=index&json=1&pid=$PAGE&tags=$TAG" | jq -r '.post')
+			JSON_URL+=$(curl -s "https://danbooru.donmai.us/posts.json?page=$PAGE&limit=200&tags=$TAG" | jq -r '.[].id')
+			if ((PAGE < TAG_PAGES)); then
+				JSON_URL+=$'\n'
+			fi
 		fi
 		sleep "$DELAY"
 	done
-
 	# NESTED LOOP FOR IMAGES IN THIS TAG
-	echo "$JSON" | jq -c '.[]' | while read -r i; do
-		#echo $i | jq -r '."id"'
-		FILE_DATE=$(echo "$i" | jq -r '."created_at"')
-		FILE_URL=$(echo "$i" | jq -r '."file_url"')
-		FILE_MD5=$(echo "$i" | jq -r '."md5"')
-		FILE_TAGS=$(echo "$i" | jq -r '."tags"' | sed 's/\ /,/g')
-		FILE=$(echo "$FILE_URL" | sed 's/\// /g' | awk '{print $NF}')
-
+	echo "$JSON_URL" | while read -r i; do
+		if $USE_TOR; then
+			JSON=$(torsocks curl -s "https://danbooru.donmai.us/posts/$i.json")
+		else
+			JSON=$(curl -s "https://danbooru.donmai.us/posts/$i.json")
+		fi
+		# STORE FILE URL AND TAGS INTO VARIABLES
+		FILE_DATE=$(echo "$JSON" | jq -r '."created_at"')
+		FILE_URL=$(echo "$JSON" | jq -r '."file_url"')
+		FILE_TAGS=$(echo "$JSON"  | jq -r '."tag_string"' | sed 's/\ /,/g')
+		FILE_MD5=$(echo "$JSON" | jq -r '.md5')
+		FILE_EXT=$(echo "$JSON" | jq -r '.file_ext')
+		FILE="$FILE_MD5.$FILE_EXT"
 		if $RESUME; then
 			if [[ -f "$FILE" ]]; then
 				echo "$FILE exists."
 				exit
 			fi
 		fi
-
 		# DOWNLOAD FILE
 		if $USE_TOR; then
 			torsocks curl -O -J "$FILE_URL"
@@ -97,6 +106,7 @@ for TAG in "${TAGS[@]}"; do
 		# ADD TAGS TO NEW IMAGE
 		setfattr -n user.xdg.tags -v "$FILE_TAGS" "$FILE"
 		setfattr --name=user.checksum --value="$FILE_MD5" "$FILE"
+		# SET TIME TO TIME UPLOADED
 		touch -d "$FILE_DATE" "$FILE"
 		# DELAY BEFORE NEXT FETCH
 		sleep "$DELAY"